Implement vmarko suspend checkss.
This is an optimization for the ARM and AARCH64 suspend checks
generated by the compiler.  It is based on an idea by vmarko
and uses a combination of a reserved register (r4 on ARM and
x19 on aarch64) and a single load instruction to perform
the suspend check.

It gives a 2.5% reduction in size and a 1.9%
speed regression on our standard benchmarks.  Yes, that's
a slow-down in speed but a significant reduction in
image size.  We also get a significant decrease in the
suspend check latency (500:1) resulting in more accurate
profiles.

Change-Id: If35ac79850dc3cea47a2d99f852c762d62f31d3e