ART: Fix evaluation of materialized conditions.
Forcing HCondition::NeedsMaterialization() and running the
compiler/optimizing/codegen_test.cc would highlight improper evaluation of
materialized conditions.
A non-zero materialized condition means that the condition was met, so the 'if'
should branch to the 'true' target.
On x86 and x86_64, the evaluation should look only at a byte, as the setcc
instruction only modifies the lower byte of the target register.