ART: Remove invalid dcheck from SplitBlock
The SplitBlock is a public interface from MIRGraph to be able to create two
blocks from one. This interface can be called not just during the control
flow graph building phase, but later. At that later point, extended
instructions may have already been inserted. The DCHECK is invalid because of
this reason.

For example, SpecialMethodInliner can insert kMirOpNullCheck and kMirOpNop,
 and then another pass tries to split block that contains this. The DCHECK
will fail even though the split is semantically correct.

Change-Id: I8a66d470a84a3836b5e68575080af1cd7278c535
Signed-off-by: Razvan A Lupusoru <razvan.a.lupusoru@intel.com>