x86_64: LongToInt should be careful in case it is nop
if destination and source of long-to-int is the same physical
register on 64-bit then we do not emmit any instructions but
consider that destination is a 32-bit view of source register.
As a result high part cantains garbage.
If the destination is used later as index to array access then
this garbage is used in computation of address because address is
64-bit. For all other cases garbage is just ignored.

To be efficient we do not do a zero extension of 32 bit value to
64-bit one always to ensure that there is no garbage. Instead of
that we remember that we did a dangerous operation and adds zero
extension only for the cases where the result is used as index
for array access.

Change-Id: I9ebe8dbe5606f5dd5b5e025cd63d56cd03b69a0d
Signed-off-by: Serguei Katkov <serguei.i.katkov@intel.com>