AArch64: tuning register promotion.
Increase the promotion threshold to 2 for Arm64. This eliminates some
stp/ldp in favour of str/ldr. Modify FlushIns() to remember the
association vreg/temp-registers to avoid extra loads in the basic
block following the entry basic block.

The combination of these changes give a reduction in the size of the
compiled binaries between 0.5% and 1.0% (measured over the oat files
compiled in a typical arm64 boot to GUI).

Change-Id: Ib9b0d5100e681bcdebd1117be4cae6030d2eae5f