ART: Calculate dataflow for all blocks
Currently dataflow information is not calculated for catch blocks in order to
reduce overhead. However, this makes assumption that the lack of this
information is only preventing local optimizations to catch block. But in
reality this problematic for global optimizations.

Basically, when dataflow is not allocated for a block, then ssa is not computed
for that block either. When ssa is not computed, other optimizations which
need to know all uses of ssa register do not see it in the catch block.
Complete information is required for correctness.

Change-Id: I5c13344c18409bd172a5d9b07577b4a1353cc733
Signed-off-by: Razvan A Lupusoru <razvan.a.lupusoru@intel.com>