Fix stale root error in verifier
There was a stale root error caused by the static roots from
the reg types. These were visitied if there was an active verifier
in the method_verifiers_ but this is not always the case when a GC
is run. The fix is to always visit the static method verifier roots.

This only showed up as a bug without an image since these roots
were primitive classes and always in the image, and therefore
didn't ever need to be updated due to moving GC.

Bug: 17262039

(cherry-picked from commit 6167864e28e4e12658ebdbaf1d5239acdaf4aaa4)

Change-Id: Ic9fe8a235baac85efbdf748aeaddf98139558c4e