ART: Adds --fail-on-soft-errors flag to dex2oat.
It is useful to be able to force dex2oat to report verification
errors that will occur at runtime even if they aren't flagged
at compile-time, if we want to perform verification of files on
our host machine before transferring them to the target.
Now with dex2oat --fail-on-soft-errors, any soft errors will be
treated as hard, even at compile-time.

Change-Id: Ia1e7ce2132f9fccd86cff27378c93a2d818f7a49
Signed-off-by: Stephen Kyle <stephen.kyle@arm.com>