Handle constant conditions in x86 & x86-64 code generators.
The constant folding pass of the optimizing compiler may
replace the condition (input) of an art::HIf instruction
with a compile-time constant (an art::HConstant instruction).
This change lets x86 and x86-64 code generators handle such
constant conditions by assigning them to a register beforehand,
so that they can be used as first operand of cmpl
instructions.

Change-Id: Ib7e0a9fe347cfe8df5a4687ace296feb0907ce32