ART: fix 32-bit wide GenInlinedMinMax for unused result
As the number of registers is tight in the 32-bit wide
GenInlinedMinMax case, the lowering starts with reading VR assigned
to the result of the min/max call to compare it with VRs of src1 and
src2. But the result may be unassigned to any VR. And the SRegToVReg
call is invalid in this case.

The patch just sets the result's VR to -1 making it different from
sources' VRs, which cannot be INVALID as the min/max call arguments.

Another option would be to not generate anything at all if the
result is unassigned to a VR, but it could be unsafe for those who
expects the result in the AX:DX pair.

Change-Id: I172bf9facd62ac4df16bd9e0b11fe472e04424f5
Signed-off-by: Serguei Katkov <serguei.i.katkov@intel.com>
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>