Stop using |instance_| in Runtime::InitNativeMethods.
Runtime::InitNativeMethods is a member function so we
don't need to use |instance_| to call Runtime methods there.

I believe that this call uses |instance_| instead of |this|
because it was originally in a static Runtime::Create method.
It was originally called 'instance_->InitLibraries();' and was
moved to Runtime::Start which is a member function here:
https://googleplex-android-review.git.corp.google.com/#/c/133053/6/src/runtime.cc
and since then it was further moved and renamed but it kept using |instance_|
even though use of |this| seems more appropriate.

Change-Id: I79b9d23873272ba2f124e285602f1b2683c2f27f