X86 Long Min/Max: Avoid calling SRegToVReg with -1
It is possible that the result of a call to min/max can't be combined
with the following move.  In that case, the destination will use the
default long return value (EAX/EDX), with a s_reg_low value of -1.
A debug compiler will assert fail in that case.

Seen compiling GmsCore.apk on the host with a debug backend.

Change-Id: I8006e822e8dcb2112d86e4047bb2e3037ba6fece
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>