Deny access to local values for proxy method for debugger
Returns JDWP OPAQUE_FRAME error when accessing stack local values
(with StackFrame.GetValues and StackFrame.SetValues). This prevents
us from crashing when accessing arguments of a proxy method as long
as we don't support access to them.

Also updates FindFrameVisitor to be used in StackFrame.ThisObject so
we can remove GetThisVisitor and avoid duplicating code. The native
method case is now handled in commands StackFrame.GetValues and
StackFrame.SetValues.

Bug: 17885923
Change-Id: Ibc6c3d382cb1e5985f556ab575b71f126b1861d1