[WIP] AArch64: Ability to predict leaf method generation.
Add code to scan a MIR graph to predict whether this will be compiled
as a leaf method or not. This ability is backend-specific, as some
backends may implement one opcode as a calls to a C helper, while some
other backends may not. A backend needs to provide an implementation of
GetOpcodeCallTargets() in order to use this functionality.

Change-Id: I06c435aeef9d30c35cae941a30fd4c6531b84c44