ART: Make InToRegStorageMapper use shorties.
[00029c] Main.main:([Ljava/lang/String;)V
0000: const/4 v0, #int 1 // #1
0001: invoke-static {v0}, LMain;.doInteger:(I)V // method@0002
0004: invoke-static {v0}, LMain;.doFloat:(F)V // method@0001
0007: return-void

In this code snippet, the call to doInteger will load v0 into a FP reg,
because type inference determines v0 to be both a float and an integer
register, due to the later use as a float.  The code that determines
which physical registers to use to pass arguments is biased towards when
this .fp flag is set.

Instead of using the RegLocation type inference information, which may
be confusing and wrong, the method's shorty info should be used instead.

Change-Id: I4452714e888244fee9991d1b65b364d9873d5a31
Signed-off-by: Stephen Kyle <stephen.kyle@arm.com>