ART: Force clearing higher bits when converting long to int
The following problem description is taken from
https://android-review.googlesource.com/#/c/107261
  If destination and source of long-to-int is the same physical
  register on 64-bit then we do not emit any instructions but
  consider that destination is a 32-bit view of source register.
  As a result high part contains garbage. If the destination is
  used later as index to array access then this garbage is used
  in computation of address because address is 64-bit. For all
  other cases garbage is just ignored.

For 64-bit target this patch forces generating reg-to-reg copy
even if the src and dest are the same physical registers. This
makes the higher bits be zeroed.

Mir2Lir::GenIntToLong is created solely to cope with the limited
stack frame size of Mir2Lir::CompileDalvikInstruction.

Change-Id: I49163a7ec1fd3c22a4747be89a32a18cea4dc9aa
Signed-off-by: Serguei Katkov <serguei.i.katkov@intel.com>
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>