[WIP] AArch64: Leaf method optimisation.
For true-leaf methods we use a different register allocation scheme.
Caller-saved registers are used as additional promotion targets. This
allows reducing the size of the stack frame (and potentially avoid
allocating it), remove spills/unspills and register moves. True-leaf
methods are typically hugely simplified by this optimisation.

Note that we currently rely on setting kLeafOptimization to true.
If this is set to false, then all methods (except from special ones)
end up being non-leaf, due to calls to the suspend check handler.

TODO: make sure we can apply this optimisation even when
kLeafOptimization is set to false.

Change-Id: Id81252cfb5a4d2657f54dc136196888f03a1fc35