[WIP] JDWP: Prepare access to proxy method arguments
Moves QuickArgumentVisitor and BuildQuickArgumentVisitor classes to
a separateto module so we can use that from debugger to access proxy
method arguments.

We now have a special handling for proxy method to implement the
StackFrame.ThisObject command. However we don't add the support for
other arguments. This will be done in a later CL.

Bug: 17965861
Change-Id: I31bac3191a082a0cfff7d88fe29d76e733cddf76