[WIP] Quick: Redefine the notion of back-egdes.
Redefine a back-edge to really mean an edge to a loop head
instead of comparing instruction offsets. Generate suspend
checks also on fall-through to a loop head.

Rewrite suspend check elimination to track whether there is
an invoke on each path from the loop head to a given back
edge, instead of using domination info to look for a basic
block with invoke that must be on each path. Ignore invokes
to intrinsics and move the optimization to a its own pass.

The new loop with an intrinsic in 109-suspend-check should
prevent intrinsics-related regressions.

TODO: Consider retaining MIR_IGNORE_SUSPEND_CHECK and adding
an extra GOTO for fall-through edges with suspend check.

Bug: 18522004
Change-Id: I239184b6c6b052771a365307a579d4f19586fe5b