ART: Added Checker, a flexible pattern matching engine for writing tests
This patch adds a Python script which implements a domain-specific mini-language similar to that of
LLVM's FileCheck. It is primarily intended for writing tests for the optimizing compiler but could
be configured for other use cases too. It is implemented from scratch in order to avoid depending
on LLVM.

Checker tests are written either in Java or smali and dex2oat is invoked with a flag which dumps
the HGraph before and after each pass of the optimizing compiler. The graph dump is then compared
against assertions located in comments in the test's source file and understood by Checker. See
comments in tools/checker.py for more details about the currently supported language features.

This initial CL implementats only one type of assertion - whether the output contains lines matching
a desired pattern in the given order - but supports both plain text and regex matching and allows
for equivalency testing by matching for the outcome of a previous match. See the constant folding
tests in compiler/optimizing/test/ConstantFolding.java for examples.

Change-Id: I1ad7431b399c38dc0391ccee74d2c643ba0b0675