[WIP] Create OatFileManager class for managing oat files.
The oat file manager is used for determining whether dex2oat or patchoat is
needed, for running dex2oat or patchoat as needed to make an oat file up to
date, and to load dex files associated with a given dex location.

The introduction of the OatFileManager class is meant to clean up and
consolidate code related to the management of oat files that was duplicated
and spread across dalvik_system_DexFile.cc and class_linker.cc.

Bug: 11301553
Change-Id: I0c16027b9bae4570c2c50faa9c14f581c0cbafb8