Make Barrier robust against spurious wakeups
Fix Barrier implementation so that Wait cannot return prematurely
due to a spurious condition variable wakeup or interrupted futex call.

Document the resulting semantics of barrier.h, which are a bit surprising,
but appear to be exactly what current clients need.

Fix the test so that it actually passes with the fixed barrier.h,
and no longer tests for properties that can't be correctly satisfied.

THIS VERSION OF THE PATCH HAS NOT YET BEEN TESTED.  Art/runtime
compiles, but the rest of my current AOSP tree doesn't, for unrelated
reasons.

Change-Id: I0b25f33bcd22322ba04e3951cd484843788c2bf5