Tweak and enable computed goto interpreter on the host
Enable the computed goto interpreter on the host compiled with clang.
Avoid external template specialized Execute...Impl functions, which brings
the compile time down for a re-build of the interpreter from 3 minutes to 2
minutes.
Tweak the alternate handler implementation to use a goto rather than computed
goto to dispatch to the opcode.

Change-Id: Ib53360a80713cafd0ef137258d69f77b43643bb3