runtime: Add support for SelfPatchoat
In this patch, OatFileAssistant is improved to take
into account status of the OatFile.
Status is split for more clarity into DexOptStatus
indicating the overall status of a file optimization
and OatStatus indicating the details status for the odex
or oat file.
This information is then used within the framework to
patch files in the dalvik-cache which are not out-of-date
and only need their base offset adjusted.

A new test is introduced to validate the self patching
of the oat when there is no odex.

Change-Id: Ib9a6373f98474f1242367b5285086251a9d580e5
Signed-off-by: Julien Delayen <julien.delayen@intel.com>
Signed-off-by: Guilhem Imberton <guilhem.imberton@intel.com>