runtime: Add support for preoptimized odex in data
In this patch, OatFileAssistant is improved to take
into account status of the OatFile.
This status can included the new kOatNeedsPatchoat
to perform a self relocation of the oat file.
This information is then used within the framework to
patch files in the dalvik-cache which are not out-of-date
and only need their base offset adjusted.

When preoptimizing odex at build time and storing them
in /data, do not to wipe the dalvik-cache during the
first boot in RelocateImage.
A flag file .preloaded is generated in the dalvik-cache
folder at build time for the given instruction set and
checked when trying to erase in RelocateImage.

A new test is introduced to validate the self patching
of the oat when there is no odex.

Change-Id: I9e5742193a24945549b092a69cc8a136e16a5d8c
Signed-off-by: Julien Delayen <julien.delayen@intel.com>
Signed-off-by: Guilhem Imberton <guilhem.imberton@intel.com>