Art: fix systematic error in sampling profiler.
SUSPEND_CHECK_INTERVAL (for explicit suspend checks) was 1000.
This introduced systematic errors in the sampling profiler.
In this patch this value is changed to 997, a prime number.

This improves the profiling data and drastically reduces the
probability that the countdown in the suspend register "aligns"
with specific program counters during execution. In particular,
consider the following Java code:

  while (true) {
    i = simpleMethod1(i, 2);  // Assume this does 1 suspend check.
    i = simpleMethod2(i, 1);  // Assume this does 1 suspend check.
  }

In the loop above, the value of the suspend counter which
simpleMethod1 sees has always the same parity. The same holds
for simpleMethod2. If SUSPEND_CHECK_INTERVAL is even, then only
one of these two methods will ever call the suspend check handler.

Change-Id: Idf4ebfdf7ecdef5a57ed0ca6fa05641e062e59de